serv_er/backend/requirements.txt
Flask
uWSGI
psycopg2-binary
Flask-SQLAlchemy
requests
pytest
alembic
gunicorn